# [PERF] mail: hundreds of chat bubbles load faster