Before this commit, when user had more about hundreds of chat bubbles, it was very slow to render them.

With 100 chat bubbles, it took 2.5seconds.

This happens because when there are lots of chat bubbles, most of them are actually hidden. The computation of hidden chat bubbles was very inefficient: it was recomputing the list each time a new chat bubble was opened or folded.

There's no need for this extra list: it can be deduced with a slice on folded chat windows. This is the core change of this commit that improves drastically performance to 400ms.

This commit scraps also `actually` fields in chatHub, which are actually useless as they are exactly the same as the non-actually fields. This reduces the load on the recomputing even further to loads 100 chat bubbles initially in 300ms.

With 100 chat bubbles:
- Before this commit: 2.5 seconds.
- After this commit: 300ms.

opw-4127973